move tableau to separate header
