+++ integration/ROS2S/Make.package
@@ -1,4 +1,5 @@
 CEXE_headers += ros2s_type.H
+CEXE_headers += ros2s_tableau.H
 CEXE_headers += ros2s_integrator.H
 ifeq ($(USE_ALL_SDC), TRUE)
   CEXE_headers += actual_integrator_sdc.H
